Extract page parsing from MF Weather Provider.
