+++ weather/src/main/java/basicWeatherInterpolation/localizationService/LocalizationService.java
@@ -1,9 +1,6 @@
 package basicWeatherInterpolation.localizationService;
 
-import basicWeatherInterpolation.weatherProviders.MountainForecastWeatherProvider;
-import basicWeatherInterpolation.weatherProviders.PogodynkaWeatherProvider;
-import basicWeatherInterpolation.weatherProviders.ImgwWeatherProvider;
-import basicWeatherInterpolation.weatherProviders.WeatherProvider;
+import basicWeatherInterpolation.weatherProviders.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,7 +15,7 @@ public class LocalizationService {
         weatherProviders.add(new ImgwWeatherProvider("Kasprowy WIerch"));
         weatherProviders.add(new ImgwWeatherProvider("Zakopane"/*TODO: Coords of IMGW station, repeat for all stations*/));
         weatherProviders.add(new PogodynkaWeatherProvider(/*TODO: Coords of IMGW station, repeat for all stations*/));
-        weatherProviders.add(new MountainForecastWeatherProvider(/*TODO: Coords of MF station, repeat for all stations*/));
+        weatherProviders.add(new MountainForecastWeatherProvider(new MountainForecastPageParser(),PeakName.Rysy));
     }
