+++ weather/src/main/java/basicWeatherInterpolation/weatherProviders/MountainForecastPageParserTest.java
@@ -1,0 +1,15 @@
+package basicWeatherInterpolation.weatherProviders;
+
+import org.junit.jupiter.api.Test;
+
+class MountainForecastPageParserTest {
+
+
+    @Test
+    public void Test() throws Exception {
+        MountainForecastPageParser sut = new MountainForecastPageParser();
+
+    }
+
+
+}
