--- weather/src/main/java/basicWeatherInterpolation/weatherProviders/MountainForecastWeatherProviderTest.java
@@ -1,22 +1,0 @@
-package basicWeatherInterpolation.weatherProviders;
-
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.select.Elements;
-import org.junit.jupiter.api.Test;
-
-import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-class MountainForecastWeatherProviderTest {
-
-
-    @Test
-    public void Test() throws Exception {
-        MountainForecastWeatherProvider sut = new MountainForecastWeatherProvider();
-
-    }
-
-
-}
