+++ weather/src/main/java/basicWeatherInterpolation/weatherProviders/PeakName.java
@@ -1,0 +1,29 @@
+package basicWeatherInterpolation.weatherProviders;
+
+public enum PeakName{
+    Banikov("Baníkov"),
+    Baranec("Baranec"),
+    Chopok("Chopok"),
+    Derese("Derese"),
+    Dumbier("Dumbier"),
+    Gerlach("Gerlach"),
+    Giewont("Giewont"),
+    Bubalowka("Gubałówka"),
+    KasprowyWierch("Kasprowy Wierch"),
+    Koscielec("Koscielec"),
+    Krivan("Krivan"),
+    Mnich("Mnich (mountain)"),
+    OstryRohac("Ostry Rohac"),
+    Rysy("Rysy");
+
+    private String name;
+
+    PeakName(String name){
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return name;
+    }
+}
